requirements.txt

torch 
transformers 
accelerate 
bitsandbytes 
sentence-transformers 
langchain 
langchain-community 
langchain-core 
langchain-huggingface 
huggingface_hub 
datasets 
faiss-cpu 
openai-whisper 
faster-whisper 
moviepy 
pydub 
ffmpeg-python 
fastapi 
uvicorn 
python-multipart 
pydantic 
requests 
tqdm 
numpy 
pandas 
scikit-learn 
matplotlib 
jupyter 
ipywidgets 
ollama
rank-bm25

groq
yt-dlp
youtube-transcript-api